Parse zuliprc file directly
Quick and dirty parsing of the zuliprc file. This would break if the
file format changes, so hopefully it doesn't do that. The parser should
be very simple to update if that happens.

I also removed the leading `$` from the example commands, because it
made them hard to copy with the "copy" button.